+++ src/apps/ScopeChangeRequest/Components/Sidesheet/Tabs/Request/RevisionList/RevisionList.tsx
@@ -1,7 +1,7 @@
 import { tokens } from '@equinor/eds-tokens';
 import { useQuery } from 'react-query';
 import styled from 'styled-components';
-import { openNewScopeChange } from '../../../../../functions/openNewScopeChange';
+import { openNewScopeChange } from '../../../../../utils/helpers/openNewScopeChange';
 import { useScopeChangeContext } from '../../../../../hooks/context/useScopeChangeContext';
 import { scopeChangeQueries } from '../../../../../keys/queries';
 import { ScopeChangeRequest } from '../../../../../types/scopeChangeRequest';
