+++ src/apps/ScopeChangeRequest/Components/Workflow/Contributor/getContributorStatus.ts
@@ -1,5 +1,4 @@
-import { CriteriaStatus } from '../Criteria/Components/CriteriaDetail';
-import { Contributor } from '../../../types/scopeChangeRequest';
+import { Contributor, CriteriaStatus } from '../../../types/scopeChangeRequest';
 
 export function getContributorStatus(
     contributor: Contributor,
