+++ src/apps/ScopeChangeRequest/ScopeChangeRequestApp.tsx
@@ -1,31 +1,14 @@
 import { ClientApi } from '@equinor/lighthouse-portal-client';
-import { setupWorkspaceSidesheet } from '../../Core/WorkSpace/src/WorkSpaceApi/Functions/setupWorkspaceSidesheet';
-import { SidesheetWrapper } from './Components/Sidesheet/SidesheetWrapper/SidesheetWrapper';
 import { ScopeChangeRequest } from './types/scopeChangeRequest';
 import {
     statusBarConfig,
     dataSource,
-    idResolver,
     prefetchQueriesOptions,
     filterConfig,
     gardenConfig,
     tableConfig,
+    sidesheetCreator,
 } from './utils/config';
-
-const sidesheetCreator = setupWorkspaceSidesheet<ScopeChangeRequest, 'change'>({
-    id: 'change',
-    color: '#7B3A96',
-    component: SidesheetWrapper,
-    props: {
-        objectIdentifier: 'id',
-        parentApp: 'change',
-        function: idResolver,
-    },
-});
-
-export const changeSideSheetWidgetManifest = sidesheetCreator('SidesheetManifest');
-export const changeSideSheetWidgetComponent = sidesheetCreator('SidesheetComponentManifest');
-export const changeResolverFunction = sidesheetCreator('ResolverFunction');
 
 export function setup(appApi: ClientApi): void {
     appApi
