+++ src/apps/ScopeChangeRequest/hooks/mutations/useRequestMutations.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from 'react-query';
 import { patchScopeChange, postScopeChange, uploadAttachment } from '../../api/ScopeChange/Request';
-import { openNewScopeChange } from '../../functions/openNewScopeChange';
+import { openNewScopeChange } from '../../utils/helpers/openNewScopeChange';
 import { scopeChangeQueryKeys } from '../../keys/scopeChangeQueryKeys';
 import { ScopeChangeCreateEditModel } from '../../types/scopeChangeRequest';
