+++ src/apps/ScopeChangeRequest/utils/helpers/openNewScopeChange.ts
@@ -1,7 +1,7 @@
 import { openSidesheet } from '@equinor/sidesheet';
-import { getScopeChangeById } from '../api/ScopeChange/Request';
-import { SidesheetWrapper } from '../Components/Sidesheet/SidesheetWrapper/SidesheetWrapper';
-import { changeSideSheetWidgetManifest } from '../ScopeChangeRequestApp';
+import { getScopeChangeById } from '../../api/ScopeChange/Request';
+import { SidesheetWrapper } from '../../Components/Sidesheet/SidesheetWrapper/SidesheetWrapper';
+import { changeSideSheetWidgetManifest } from '../config/sidesheetConfig';
 
 export async function openNewScopeChange(scopeChangeId: string): Promise<void> {
     openSidesheet(
